Rename des -> signature
