+++ src/lib.rs
@@ -10,7 +10,6 @@ pub use gd32vf103_pac as pac;
 use embedded_hal as hal;
 
 pub mod afio;
-pub mod des;
 pub mod exmc;
 pub mod gpio;
 pub mod i2c;
@@ -18,6 +17,7 @@ pub mod rcu;
 pub mod backup_domain;
 pub mod prelude;
 pub mod serial;
+pub mod signature;
 pub mod spi;
 pub mod time;
 pub mod timer;
